1. ModuleDrainageNetwork - verified if discharge negative to avoid negative volumes and dt cutting 2. ModuleDrainageNetwork - now it is possible to have multiple outlets instead of one previously using a list of flows. OPENMI has to be reviewed to be consistent with this 3. Vegetation - modified the maximum root depth verification. the old version was limiting the values to the minimum soil depth observed in the grid per vegetation ID.
git-svn-id: https://mohid.svn.codeplex.com/svn@85740 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ Software/MOHIDLand/ModuleBasin.F90
@@ -505,7 +505,7 @@ Module ModuleBasin
         real, dimension(:), pointer                 :: TimeSeriesBuffer3 !Properties Balance
         real, dimension(:), pointer                 :: TimeSeriesBuffer4 !Water Fluxes
         real, dimension(:), pointer                 :: TimeSeriesBuffer5 !Properties Fluxes
-        real(8), dimension(:), pointer              :: BWBBuffer !buffer to be used for Basin Water Balance
+        real, dimension(:), pointer                 :: BWBBuffer !buffer to be used for Basin Water Balance
         
         !Basin Water Balance
         type (T_BasinWaterBalance)                  :: BWB
